requirements.txt
asgiref==3.2.3
blessings==1.7
bpython==0.18
cachetools==4.1.0
certifi==2019.11.28
chardet==3.0.4
curtsies==0.3.1
dj-database-url==0.5.0
Django==3.0.3
django-cors-headers==3.2.1
django-filter==2.2.0
django-storages==1.9.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
google-api-core==1.17.0
google-auth==1.14.3
google-cloud-core==1.3.0
google-cloud-storage==1.28.1
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
greenlet==0.4.15
gunicorn==20.0.4
idna==2.9
Markdown==3.2.1
Pillow==7.1.2
protobuf==3.11.3
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.5.2
PyJWT==1.7.1
pytz==2019.3
requests==2.23.0
rsa==4.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
wcwidth==0.1.8
